requirements-batik.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
tensorflow==2.19.0
pillow==10.0.0
pandas==2.0.3
numpy==1.26.4
scikit-learn==1.3.0
umap-learn==0.5.3
opencv-python-headless==4.8.1.78
psutil==5.9.5
joblib==1.3.2
python-dotenv==1.0.0
